refactor(096-W2): kilo/opencode 提取器合并为 StepProtocolExtractor——step 协议族唯一事实源
- 093-B1 已把两家事件结构体收敛为 StepAgentEvent 别名，本 commit 完成实现层合并：
  新建 impls/step_protocol.rs（new(executor_name) 参数化唯一身份差异），
  kilo.rs/opencode.rs 降级为类型别名壳（引用路径不抖）
- text/agent 与 tool-use 两分支的 ToolCall 构建收敛为 tool_call_event  helper
- 行为差异（093-B1 既定语义对齐，声明）：tool input 序列化从「extra 允许覆盖
  command/description」的旧 ext trait 切换为 to_full_json_value 防覆盖语义；
  StepAgentToolInput 新增 to_full_json_value（Value 形态），to_full_json 委托之
- 测试：kilo 7 + opencode 5 合并去重为 11 项协议行为锚点；全量 1670 过、clippy 零告警
